refactor!: move Hand struct into separate module
This is a step toward keeping the `lib.rs` module tidy.
